feat: 新增用户注册接口

+++ easy-pan-back/src/main/java/com/easy_pan/back/model/bo/EmailVerifyCodeBO.java
@@ -7,6 +7,6 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class EmailVerifyCodeBO {
-    private Integer code;
-    private Integer status;
+    private long code;
+    private int status;
 }
